KSENROLL-13104 Create tool to remove .mpx files from project history
Removing the .sql files is non-trivial and complicated but the .mpx files are
included in the published jars.

It should be possible to purge them from the repository.

Adding a new cleaner that will let us test how much space can be reclaimed if
we remove all of the .mpx files.

git-svn-id: https://svn.kuali.org/repos/student/tools/svn-history-tools/trunk@73642 0ac02d80-fb3f-11dc-a574-c57244a850eb